% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_novelty_to_test_data.R
\name{add_novelty_to_test_data}
\alias{add_novelty_to_test_data}
\title{Relabel the true test annotations based on a reference library used to train a BayesANT model}
\usage{
add_novelty_to_test_data(data, test_data)
}
\arguments{
\item{data}{An object of class \code{c('data.frame', 'BayesANT.data')}}

\item{test_data}{An object of class \code{c('data.frame', 'BayesANT.data')}}
}
\value{
An object of class \code{c('data.frame', 'BayesANT.data')}
}
\description{
Relabel the true test annotations based on a reference library used to train a BayesANT model
}
