% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesANT.R
\name{BayesANT}
\alias{BayesANT}
\title{Call the BayesANT Classifier.}
\usage{
BayesANT(
  data,
  typeseq = "aligned",
  type_location = "single",
  kmers = 5,
  newtaxa = TRUE,
  usegap = FALSE,
  adjust_Kmer_length = TRUE,
  save_nucl = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{data}{An object of class \code{c('data.frame', 'BayesANT.data')}. Needs to be loaded with the function \code{read.BayesANT.data}.}

\item{typeseq}{Type of sequences used to train the classifier. Options are \code{'not aligned'} and \code{'aligned'}. Default is \code{typeseq = 'aligned'}}

\item{type_location}{How to model the loci in an aligned set of DNA sequences. Valid only if \code{typeseq = 'aligned'}.
Options are \code{'single'}, which refers to the single-location multinomial kernel, and  \code{'pairs'}, which is the 2-Product multinomial kernel. Default is \code{type_location = 'single'}}

\item{kmers}{How many kmers to take for the k-mer decomposition. Valid only if \code{typeseq = 'not aligned'}. Default is  \code{kmers = 5}}

\item{newtaxa}{Include new taxa when constructing the classifier. Default is \code{newtaxa = TRUE}}

\item{usegap}{Whether to include the alignment gap \code{"-"} among the nucleotides in the aligned multinomial kernel. Default is \code{usegap = FALSE}}

\item{adjust_Kmer_length}{Adjust the likelihood contribution of a kmer decomposition based on its length. Valid for the option  \code{typeseq ='not aligned'}.}

\item{save_nucl}{Save the nucleotide counts and the hyperparameters of the model in a list. Default is \code{save_nucl = FALSE}. Setting it to \code{TRUE} might severely increase the storage requirements, so be careful.}

\item{verbose}{Monitor the state of the algorithm. Default is \code{verbose = TRUE}}
}
\value{
An object of class \code{BayesANT}
}
\description{
Call the BayesANT Classifier.
}
