% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesANT_amino.R
\name{BayesANT_amino}
\alias{BayesANT_amino}
\title{Main function to call the BayesANT (BAYESiAn Nonparametric Taxonomic)
classifier. The function construct an object of class \code{'BayesANT'},
which can be used for predictions.}
\usage{
BayesANT_amino(
  data,
  typeseq = "aligned",
  type_location = "single",
  kmers = 5,
  newtaxa = TRUE,
  usegap = FALSE,
  save_nucl = FALSE,
  amino = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{data}{An object of class \code{c('data.frame', 'BayesANT.data')}
containing the training library. Needs to be loaded with the
function \code{read.BayesANT.data}.}

\item{typeseq}{Format of sequences used to train the classifier.
Options are \code{'not aligned'} and \code{'aligned'}.
Default is \code{typeseq = 'aligned'}. In this last case,
the function returns an error if the sequences are not of the
same length.}

\item{type_location}{How to model the loci in an aligned set of DNA
sequences. Valid only if \code{typeseq = 'aligned'}.
Options are \code{'single'}, which refers to the
single-location multinomial kernel, and  \code{'pairs'},
which is the 2-mer multinomial kernel.  Default is
\code{type_location = 'single'}}

\item{kmers}{Length of a substring under a k-mer decomposition.
Valid only if \code{typeseq = 'not aligned'}.
Default is  \code{kmers = 5}, which is also the recommended
choice. The maximum choice allowed is \code{kmers = 8}}

\item{newtaxa}{Whether to account for new taxa when constructing the
classifier. Default is \code{newtaxa = TRUE}. If
\code{newtaxa = FALSE},no potential unobserved branches are
included in the taxonomy.}

\item{usegap}{Whether to include the alignment gap \code{"-"} among the
nucleotides in the aligned multinomial kernel. Default is
\code{usegap = FALSE}}

\item{save_nucl}{Save the nucleotide counts and the hyperparameters of
the model in a list. Default is \code{save_nucl = FALSE}.
Setting it to \code{TRUE} might be heavy to store, so use
only if strictly needed.}

\item{verbose}{Monitor the steps adopted to train the algorithm.
Default is \code{verbose = TRUE}.}
}
\value{
An object of class \code{BayesANT}. We return a \code{list}
        containing the following quantities:
        \itemize{
         \item{data}{Dataset used for training.},
         \item{data_missing}{Dataset containing sequences with missing
            annotations.}
         \item{missing_taxa}{Indeces of the sequences with missing values.}
         \item{typeseq}{Type of sequences used to train the classifier.}
         \item{type_location}{Type of location to train the classifier.}
         \item{newtaxa}{Whether new taxa are included in the classification.}
         \item{level_names}{Names of taxonomic ranks}
         \item{nucl}{Nucleotides detected.}
         \item{kmers}{Number of kmers selected to build the classifier.}
         \item{ParameterMatrix}{Matrix that stores model parameters.}
         \item{Nucl_counts}{List of counts of the nucleotides at every leaf.}
         \item{PYpars}{Estimated Pitman-yor parameters.}
         \item{Priorprobs}{Prior probabilities selected for the model.}
         \item{hyperparameters}{List of model hyperparameters.}
         \item{leaves}{Name of the taxonomic leaves.}
         \item{sequences_length}{Length of each sequence in the data.}
        }
}
\description{
Main function to call the BayesANT (BAYESiAn Nonparametric Taxonomic)
classifier. The function construct an object of class \code{'BayesANT'},
which can be used for predictions.
}
