% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitPY_tree.R
\name{fitPY_tree}
\alias{fitPY_tree}
\title{Fit the Pitman-Yor parameters}
\usage{
fitPY_tree(tree)
}
\arguments{
\item{tree}{A tree that comes out of the simulate_tree}
}
\value{

}
\description{
Fit the Pitman-Yor parameters
}
