% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_hyperparameters.R
\name{get_proportions}
\alias{get_proportions}
\title{Extract the proportions of the given Nucleotide counts}
\usage{
get_proportions(Nucl_counts, eps = 0, verbose = FALSE)
}
\arguments{
\item{Nucl_counts}{nulcelotide counts list}

\item{eps}{small tweak for the zero counts}
}
\value{

}
\description{
Extract the proportions of the given Nucleotide counts
}
