% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.BayesANT.R
\name{predict.BayesANT}
\alias{predict.BayesANT}
\title{Predict the taxonomic annotation for query DNA sequences via a BayesANT model.}
\usage{
\method{predict}{BayesANT}(
  object,
  DNA,
  rho = 0.1,
  return_probs = FALSE,
  n_top_taxa = 5,
  cores = 1,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{an object of class \code{BayesANT}}

\item{DNA}{DNA sequence to predict. Consider loading it with the function
\code{read.BayesANT.testDNA}}

\item{rho}{Temperature parameter to re-calibrate the leaf probabilities.
Default set to \code{rho = 0.1}.}

\item{return_probs}{Whether to return the first \code{n_top_taxa} leafs that have
the highest probability. Default is
\code{return_probs = FALSE}.}

\item{n_top_taxa}{Number of leafs to return for the prediction. Default is
\code{n_top_taxa = NULL}, which reports the full taxonomic tree.
Valid only if \code{return_probs = TRUE}.}

\item{cores}{Optional. Specify the number of cores for predicting DNA
sequences in parallel. Relies on a combination of the packages
\code{foreach} and \code{doParallel}. Default is
\code{cores = 1}, which corresponds to a standard \code{for}
loop.}

\item{verbose}{Monitor the number of sequences predicted.}

\item{...}{Additional parameters}
}
\value{
An object of class \code{data.frame} or \code{list}.
}
\description{
Predict the taxonomic annotation for query DNA sequences via a BayesANT model.
}
