% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.BayesANT.testDNA.R
\name{read.BayesANT.testDNA}
\alias{read.BayesANT.testDNA}
\title{Load query DNA sequences to be used by the function \code{predict.BayesANT}}
\usage{
read.BayesANT.testDNA(fasta.file, ...)
}
\arguments{
\item{fasta.file}{File in \code{.fasta} or \code{.fas}
format containing the DNA sequences.}

\item{...}{Additional parameters.}
}
\value{
A named vector of strings
}
\description{
Load query DNA sequences to be used by the function \code{predict.BayesANT}
}
